# lenexus/lib/api-spec/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/guides": {
      "get": {"operationId":"listGuides","responses":{"200":{"description":"Guides publics de la communauté","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommunityGuide"}}}}}}},
      "post": {"operationId":"createGuide","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuideInput"}}}},"responses":{"201":{"description":"Guide publié directement","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityGuide"}}}}}}
    },
    "/guides/{guideId}": {
      "parameters":[{"name":"guideId","in":"path","required":true,"schema":{"type":"string"}}],
      "get":{"operationId":"getGuide","responses":{"200":{"description":"Guide et transcription complète du fil Discord","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityGuide"}}}}}},
      "put":{"operationId":"updateGuide","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuideInput"}}}},"responses":{"200":{"description":"Guide modifié sans altérer les messages Discord importés","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommunityGuide"}}}}}},
      "delete":{"operationId":"deleteGuide","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuideRevision"}}}},"responses":{"200":{"description":"Guide retiré du site uniquement","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuideStatus"}}}}}}
    },
    "/guide-uploads/request-url": {
      "post":{"operationId":"requestGuideImage","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuideImageRequest"}}}},"responses":{"200":{"description":"Envoi direct d'une image dans le stockage durable","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuideImageUpload"}}}}}}
    },
    "/guide-uploads/publish": {
      "post":{"operationId":"publishGuideImage","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuideImagePath"}}}},"responses":{"200":{"description":"Image vérifiée et disponible pour un guide","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GuideImagePublished"}}}}}}
    },
    "/dino-catalogue": {
      "get": {"operationId":"listDinoSpecies","parameters":[{"name":"all","in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":"Catalogue lié aux maps et mods du site","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DinoSpecies"}}}}}}},
      "post":{"operationId":"createDinoSpecies","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DinoSpeciesInput"}}}},"responses":{"201":{"description":"Espèce ou variante ajoutée par un administrateur ou propriétaire du site","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DinoSpecies"}}}}}}
    },
    "/dino-catalogue/{speciesId}": {
      "parameters":[{"name":"speciesId","in":"path","required":true,"schema":{"type":"integer"}}],
      "put":{"operationId":"updateDinoSpecies","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DinoSpeciesInput"}}}},"responses":{"200":{"description":"Créature modifiée","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DinoSpecies"}}}}}}
    },
    "/tribes/{id}/dino-records": {
      "parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],
      "get":{"operationId":"listDinoRecords","responses":{"200":{"description":"Records privés de la tribu","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DinoRecord"}}}}}}}
    },
    "/tribes/{id}/dino-records/{speciesId}": {
      "parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}},{"name":"speciesId","in":"path","required":true,"schema":{"type":"integer"}}],
      "put":{"operationId":"saveDinoRecord","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DinoRecordInput"}}}},"responses":{"200":{"description":"Records confirmés","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DinoRecord"}}}}}},
      "delete":{"operationId":"deleteDinoRecord","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DinoRecordDelete"}}}},"responses":{"200":{"description":"Fiche supprimée après confirmation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DinoDeleteStatus"}}}}}}
    },
    "/tribes/{id}/dino-photo": {
      "parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],
      "post":{"operationId":"readDinoPhoto","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DinoPhotoInput"}}}},"responses":{"200":{"description":"Lecture à vérifier, sans enregistrement automatique","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DinoPhotoResult"}}}}}}
    },
    "/tribes/{id}/dino-history": {
      "parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],
      "get":{"operationId":"listDinoHistory","responses":{"200":{"description":"Les 100 dernières modifications","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DinoRecordHistory"}}}}}}}
    },
...
    "/site-maps/{slug}": {
      "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[a-z0-9-]+$", "maxLength": 80}}],
      "get": {
        "operationId": "getSiteMap", "tags": ["tribes"], "summary": "Consulter une map",
        "responses": {"200": {"description": "Map", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SiteMap"}}}}, "404": {"description": "Map introuvable"}}
      },
      "put": {
        "operationId": "updateSiteMap", "tags": ["tribes"], "summary": "Modifier une map en tant que propriétaire du site",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SiteMapInput"}}}},
        "responses": {
          "200": {"description": "Map enregistrée", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SiteMap"}}}},
          "400": {"description": "Formulaire invalide"}, "401": {"description": "Connexion requise"}, "403": {"description": "Propriétaire requis"}, "404": {"description": "Map introuvable"}, "409": {"description": "Map modifiée par un autre propriétaire"}
        }
      }
    },
    "/site-maps/{slug}/settlements": {
      "parameters": [{"name":"slug","in":"path","required":true,"schema":{"type":"string","pattern":"^[a-z0-9-]+$","maxLength":80}}],
      "get": {"operationId":"getMapSettlements","tags":["tribes"],"summary":"Tribus dont la base principale est sur cette map",
        "responses":{"200":{"description":"Implantations déclarées et carte","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MapSettlements"}}}},"404":{"description":"Map introuvable"}}}
    },
    "/site-maps/{slug}/cartography": {
      "parameters": [{"name":"slug","in":"path","required":true,"schema":{"type":"string","pattern":"^[a-z0-9-]+$","maxLength":80}}],
      "put":{"operationId":"saveMapCartography","tags":["tribes"],"summary":"Enregistrer la carte officielle en tant que propriétaire",
        "requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MapCartographyInput"}}}},
        "responses":{"200":{"description":"Carte enregistrée","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MapCartography"}}}},"400":{"description":"Image invalide"},"401":{"description":"Connexion requise"},"403":{"description":"Propriétaire requis"},"409":{"description":"Carte modifiée entre-temps"}}}
    },
...
    "/site-branding": {
      "get": {
        "operationId": "getSiteBranding",
        "tags": [
          "tribes"
        ],
        "responses": {
          "200": {
            "description": "Site identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteBranding"
                }
              }
            }
          },
          "400": {
            "description": "Invalid image or input"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Site owner required"
          },
          "503": {
            "description": "Storage temporarily unavailable"
          }
        }
      },
      "patch": {
        "operationId": "updateSiteBranding",
        "tags": [
          "tribes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteBrandingInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Site identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteBranding"
                }
              }
            }
          },
          "400": {
            "description": "Invalid image or input"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Site owner required"
          },
          "503": {
            "description": "Storage temporarily unavailable"
          }
        }
      }
    },
    "/site-branding/uploads": {
      "post": {
        "operationId": "requestBrandingUpload",
        "tags": [
          "tribes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandingUploadInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed direct-to-storage upload URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingUpload"
                }
              }
            }
          },
          "400": {
            "description": "Invalid image or input"
          },
          "401": {
            "description": "Sign in required"
          },
          "403": {
            "description": "Site owner required"
          },
          "429": {
            "description": "Too many uploads"
          },
          "503": {
            "description": "Storage temporarily unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GuideMedia": {
        "type":"object","required":["url","name","contentType"],
        "properties":{"url":{"type":"string"},"name":{"type":"string"},"contentType":{"type":"string"}}
      },
      "GuideEmbed": {
        "type":"object","required":["title","description","url","image","fields"],
        "properties":{"title":{"type":"string"},"description":{"type":"string"},"url":{"type":"string"},"image":{"type":"string"},"fields":{"type":"array","items":{"type":"object","required":["name","value"],"properties":{"name":{"type":"string"},"value":{"type":"string"}}}}}
      },
      "GuideMessage": {
        "type":"object","required":["id","authorId","authorName","content","createdAt","media","embeds","system"],
        "properties":{"id":{"type":"string"},"authorId":{"type":"string"},"authorName":{"type":"string"},"content":{"type":"string"},"createdAt":{"type":"string"},"media":{"type":"array","items":{"$ref":"#/components/schemas/GuideMedia"}},"embeds":{"type":"array","items":{"$ref":"#/components/schemas/GuideEmbed"}},"system":{"type":"boolean"}}
      },
      "CommunityGuide": {
        "type":"object","required":["id","title","content","authorId","authorName","images","messages","sourceUrl","source","createdAt","updatedAt","revision","canEdit","starterMissing"],
        "properties":{"id":{"type":"string"},"title":{"type":"string"},"content":{"type":"string"},"authorId":{"type":"string"},"authorName":{"type":"string"},"images":{"type":"array","items":{"type":"string"}},"messages":{"type":"array","items":{"$ref":"#/components/schemas/GuideMessage"}},"sourceUrl":{"type":"string"},"source":{"type":"string","enum":["discord","site"]},"createdAt":{"type":"string"},"updatedAt":{"type":"string"},"revision":{"type":"integer"},"canEdit":{"type":"boolean"},"starterMissing":{"type":"boolean"}}
      },
      "GuideInput": {
        "type":"object","required":["title","content","images","revision"],
        "properties":{"title":{"type":"string","minLength":1,"maxLength":180},"content":{"type":"string","maxLength":20000},"images":{"type":"array","maxItems":16,"uniqueItems":true,"items":{"type":"string","maxLength":256}},"revision":{"type":"integer","minimum":0}}
      },
      "GuideRevision":{"type":"object","required":["revision"],"properties":{"revision":{"type":"integer","minimum":1}}},
      "GuideStatus":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}}},
      "GuideImageRequest":{"type":"object","required":["size","contentType"],"properties":{"size":{"type":"integer","minimum":1,"maximum":10485760},"contentType":{"type":"string","enum":["image/png","image/jpeg","image/webp"]}}},
      "GuideImagePath":{"type":"object","required":["objectPath"],"properties":{"objectPath":{"type":"string","maxLength":256}}},
      "GuideImageUpload":{"type":"object","required":["objectPath","uploadURL"],"properties":{"objectPath":{"type":"string"},"uploadURL":{"type":"string"}}},
      "GuideImagePublished":{"type":"object","required":["url"],"properties":{"url":{"type":"string"}}},
      "DinoDeleteStatus":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}}},
      "DinoStats":{"type":"object","additionalProperties":false,"required":["health","stamina","oxygen","food","weight","melee","speed"],"properties":{
        "health":{"type":["integer","null"],"minimum":0,"maximum":65535},"stamina":{"type":["integer","null"],"minimum":0,"maximum":65535},"oxygen":{"type":["integer","null"],"minimum":0,"maximum":65535},"food":{"type":["integer","null"],"minimum":0,"maximum":65535},"weight":{"type":["integer","null"],"minimum":0,"maximum":65535},"melee":{"type":["integer","null"],"minimum":0,"maximum":65535},"speed":{"type":["integer","null"],"minimum":0,"maximum":65535}
      }},
      "DinoSpecies":{"type":"object","required":["id","name","origin","modId","maps","sourceUrl","active","revision"],"properties":{"id":{"type":"integer"},"name":{"type":"string"},"origin":{"type":"string","enum":["official","mod"]},"modId":{"type":["integer","null"]},"maps":{"type":"array","items":{"type":"string"}},"sourceUrl":{"type":"string"},"active":{"type":"boolean"},"revision":{"type":"integer"}}},
      "DinoSpeciesInput":{"type":"object","required":["name","origin","modId","maps","sourceUrl","active","revision"],"properties":{"name":{"type":"string","minLength":1,"maxLength":120},"origin":{"type":"string","enum":["official","mod"]},"modId":{"type":["integer","null"]},"maps":{"type":"array","items":{"type":"string"},"maxItems":50},"sourceUrl":{"type":"string","maxLength":600},"active":{"type":"boolean"},"revision":{"type":"integer","minimum":0}}},
      "DinoRecord":{"type":"object","required":["species","tribeId","stats","revision","updatedAt","note","source"],"properties":{"species":{"$ref":"#/components/schemas/DinoSpecies"},"tribeId":{"type":"integer"},"stats":{"$ref":"#/components/schemas/DinoStats"},"revision":{"type":"integer"},"updatedAt":{"type":"string"},"note":{"type":"string"},"source":{"type":"string","enum":["manual","photo"]}}},
      "DinoRecordInput":{"type":"object","required":["stats","revision","confirmed","mode","note","source"],"properties":{"stats":{"$ref":"#/components/schemas/DinoStats"},"revision":{"type":"integer","minimum":0},"confirmed":{"type":"boolean"},"mode":{"type":"string","enum":["best","replace"]},"note":{"type":"string","maxLength":1000},"source":{"type":"string","enum":["manual","photo"]}}},
      "DinoRecordDelete":{"type":"object","required":["revision","confirmed"],"properties":{"revision":{"type":"integer","minimum":1},"confirmed":{"type":"boolean"}}},
      "DinoPhotoInput":{"type":"object","required":["image"],"properties":{"image":{"type":"string","maxLength":8500000}}},
      "DinoPhotoResult":{"type":"object","required":["speciesName","stats","warnings"],"properties":{"speciesName":{"type":["string","null"]},"stats":{"$ref":"#/components/schemas/DinoStats"},"warnings":{"type":"array","items":{"type":"string"}}}},
      "DinoRecordHistory":{"type":"object","required":["id","speciesName","action","before","after","createdAt"],"properties":{"id":{"type":"integer"},"speciesName":{"type":"string"},"action":{"type":"string"},"before":{"oneOf":[{"$ref":"#/components/schemas/DinoStats"},{"type":"null"}]},"after":{"oneOf":[{"$ref":"#/components/schemas/DinoStats"},{"type":"null"}]},"createdAt":{"type":"string"}}},
...
      "SiteBranding": {
        "type": "object",
        "required": [
          "logoUrl",
          "coverUrl",
          "logoPath",
          "coverPath",
          "updatedAt"
        ],
        "properties": {
          "logoUrl": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string"
          },
          "logoPath": {
            "type": "string",
            "nullable": true
          },
          "coverPath": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "SiteBrandingInput": {
        "type": "object",
        "properties": {
          "logoPath": {
            "type": "string",
            "nullable": true,
            "pattern": "^/objects/uploads/[a-f0-9-]{36}$"
          },
          "coverPath": {
            "type": "string",
            "nullable": true,
            "pattern": "^/objects/uploads/[a-f0-9-]{36}$"
          }
        }
      },
      "BrandingUploadInput": {
        "type": "object",
        "required": [
          "kind",
          "name",
          "size",
          "contentType"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "logo",
              "cover"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10485760
          },
          "contentType": {
            "type": "string",
            "enum": [
              "image/png",
              "image/jpeg",
              "image/webp"
            ]
          }
        }
      },
...
      "BrandingUpload": {
        "type": "object",
        "required": [
          "uploadURL",
          "objectPath"
        ],
        "properties": {
          "uploadURL": {
            "type": "string"
          },
          "objectPath": {
            "type": "string"
          }
        }
      },
      "MapCartographyInput":{"type":"object","required":["image","revision"],"properties":{"image":{"type":"string","maxLength":2048},"revision":{"type":"integer","minimum":0}}},
      "MapCartography":{"type":"object","required":["image","revision"],"properties":{"image":{"type":"string"},"revision":{"type":"integer","minimum":0}}},
      "MapSettlement":{"type":"object","required":["id","name","coords","latitude","longitude"],"properties":{"id":{"type":"integer"},"name":{"type":"string"},"coords":{"type":"string"},"latitude":{"type":["number","null"],"minimum":0,"maximum":100},"longitude":{"type":["number","null"],"minimum":0,"maximum":100}}},
      "MapSettlements":{"type":"object","required":["count","tribes","cartography"],"properties":{"count":{"type":"integer","minimum":0},"tribes":{"type":"array","items":{"$ref":"#/components/schemas/MapSettlement"}},"cartography":{"$ref":"#/components/schemas/MapCartography"}}},
      "SiteMapInput": {
        "type": "object",
        "required": ["name","kind","image","size","biomes","difficulty","transfers","specials","blurb","description","creatures","zones","rules","gallery","revision"],
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 80},
          "kind": {"type": "string", "enum": ["main","premium","event"]},
          "image": {"type": "string", "maxLength": 2048},
          "size": {"type": "string", "maxLength": 200},
          "biomes": {"type": "string", "maxLength": 400},
          "difficulty": {"type": "integer", "minimum": 1, "maximum": 5},
          "transfers": {"type": "string", "maxLength": 400},
          "specials": {"type": "string", "maxLength": 400},
          "resourcesUrl": {"type": "string", "maxLength": 2048, "description": "Lien facultatif de localisation des ressources ; vide pour retirer le lien"},
          "dinosUrl": {"type": "string", "maxLength": 2048, "description": "Lien facultatif de localisation des dinos ; vide pour retirer le lien"},
          "blurb": {"type": "string", "maxLength": 1000},
          "description": {"type": "string", "maxLength": 10000},
          "creatures": {"type": "array", "maxItems": 100, "items": {"type": "string", "minLength": 1, "maxLength": 500}},
          "zones": {"type": "array", "maxItems": 100, "items": {"type": "string", "minLength": 1, "maxLength": 1000}},
          "rules": {"type": "array", "maxItems": 100, "items": {"type": "string", "minLength": 1, "maxLength": 1000}},
          "gallery": {"type": "array", "maxItems": 24, "items": {"type": "string", "minLength": 1, "maxLength": 2048}},
          "revision": {"type": "integer", "minimum": 0}
        }
      },
      "SiteMap": {
        "allOf": [
          {"$ref": "#/components/schemas/SiteMapInput"},
          {"type": "object", "required": ["slug","updatedAt"], "properties": {"slug": {"type": "string"}, "updatedAt": {"type": ["string","null"]}}}
        ]
      }
    }
  }
}